fix integer type propagation from num_rbm in controller_synthesis
Change int8(input_settings.rbm)*9 to double(...)*9 so that num_rbm is a
proper double. This removes the need for defensive double() casts
throughout define_Q_and_R_matrices_for_LQR and penalize_states.

+++ 05_Utils/matlab_functions/controller_synthesis.m
@@ -32,7 +32,7 @@ if use_elevator
 end
 num_control_surfaces = num_ailerons + num_elevators;
 num_ignored_last_columns = 1 + num_thrust_inputs; % Gust (1) + thrust inputs
-num_rbm = int8(input_settings.rbm)*9;
+num_rbm = double(input_settings.rbm)*9;
 num_cs=num_control_surfaces;
 %% Reduce cs inputs by assuming same deflection/join ailerons to one
 if join_cs
